+++ src/components/BasketSlider.jsx
@@ -2,21 +2,15 @@ import "./BasketSlider.css";
 import PropTypes from "prop-types";
 import { BsXLg } from "react-icons/bs";
 import { useState, useRef } from "react";
-import { useHistory } from 'react-router-dom';
 
 // Component Imports
 import BasketItem from "./BasketItem";
+import { Link } from "react-router-dom";
+import CheckOut from "../pages/CheckOut";
 
 export default function BasketSlider({ toggleBasket, basketItems, deleteFromBasket, basketTotal, changeQuantity }) {
     const [slideIn, setSlideIn] = useState(false);
     const sliderBg = useRef();
-
-    const history = useHistory();
-
-    const handleCheckoutClick = () => {
-        // Navigate to the checkout page
-        history.push('/checkout');
-      };
 
     const handleCloseSlider = (e) => {
         // If user clicks outside the slider or the X icon, close the basket
@@ -61,7 +55,9 @@ export default function BasketSlider({ toggleBasket, basketItems, deleteFromBask
                         <strong>Total: </strong>
                         <span>{basketItems.length === 0 ? "£0.00" : `£${basketTotal}`}</span>
                     </p>
-                    <button className="sliderFooter_button" onClick={handleCheckoutClick}>Checkout</button>
+                    <Link to="/checkout">
+                        <button className="sliderFooter_button" onClick={toggleBasket}>Checkout</button>
+                    </Link>
                 </div>
             </div>
         </div>
